+++ aiohttp_apigami/__init__.py
@@ -8,10 +8,12 @@ from .decorators import (
     form_schema,
     headers_schema,
     json_schema,
+    marshal_with,
     match_info_schema,
     querystring_schema,
     request_schema,
     response_schema,
+    use_kwargs,
 )
 from .middlewares import validation_middleware
 
@@ -24,11 +26,13 @@ __all__ = [
     "form_schema",
     "headers_schema",
     "json_schema",
+    "marshal_with",
     "match_info_schema",
     "querystring_schema",
     "request_schema",
     "response_schema",
     "setup_aiohttp_apispec",
+    "use_kwargs",
     "validation_middleware",
 ]