+++ aiohttp_apigami/decorators/__init__.py
@@ -8,5 +8,6 @@ from .request import (
     match_info_schema,
     querystring_schema,
     request_schema,
+    use_kwargs,
 )
-from .response import response_schema
+from .response import marshal_with, response_schema